Api Restful for CRUD

+++ src/Controller/ItemController.php
@@ -52,16 +52,23 @@ class ItemController
      * Edit item by $id
      *
      * @param int $id
-     * @return string
      */
-    public function edit(int $id): string
+    public function edit(int $id)
     {
-        $itemManager = new ItemManager();
-        $item = $itemManager->selectOneById($id);
-
-        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $item['title'] = $_POST['title'];
-            $itemManager->update($item);
+        if ($_SERVER['REQUEST_METHOD'] === 'PUT') {
+            try {
+                $itemManager = new ItemManager();
+                $item = $itemManager->selectOneById($id);
+                $json = file_get_contents('php://input');
+                $obj = json_decode($json);
+                $item['title'] = $obj->title;
+                $itemManager->update($item);
+                header('HTTP/1.1 204 resource updated successfully');
+            } catch (\Exception $e) {
+                /* var_dump should be delete in production */
+                var_dump($e->getMessage());
+                header('HTTP/1.1 500 Internal Server Error');
+            }
         }
     }
 
@@ -74,13 +81,20 @@ class ItemController
     public function add()
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-            $json = file_get_contents('php://input');
-            $obj = json_decode($json);
-            $itemManager = new ItemManager();
-            $item = [
-                'title' => $obj->title,
-            ];
-            $itemManager->insert($item);
+            try {
+                $json = file_get_contents('php://input');
+                $obj = json_decode($json);
+                $itemManager = new ItemManager();
+                $item = [
+                    'title' => $obj->title,
+                ];
+                $itemManager->insert($item);
+                header('HTTP/1.1 201 Created');
+            } catch (\Exception $e) {
+                /* var_dump should be delete in production */
+                var_dump($e->getMessage());
+                header('HTTP/1.1 500 Internal Server Error');
+            }
         }
     }
 
@@ -92,7 +106,15 @@ class ItemController
      */
     public function delete(int $id)
     {
-        $itemManager = new ItemManager();
-        $itemManager->delete($id);
+        if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
+            try {
+                $itemManager = new ItemManager();
+                $itemManager->delete($id);
+            } catch (\Exception $e) {
+                /* var_dump should be delete in production */
+                var_dump($e->getMessage());
+                header('HTTP/1.1 500 Internal Server Error');
+            }
+        }
     }
 }
